+++ lib/dia/commonapi.rb
@@ -4,6 +4,8 @@ require 'ffi'
 module Dia
   module CommonAPI
     extend FFI::Library
+    ffi_lib('sandbox')
+    
     attach_function :sandbox_init, [ :string, :int, :pointer ], :int
   end
 end